[impl]: cfg construction

+++ Code/bitset.h
@@ -2,17 +2,17 @@
 #define __CMM_BITSET_H__
 #include <stdint.h>
 typedef struct {
-  int limit;
-  int setlen;
+  uint32_t limit;
+  uint32_t setlen;
   uint64_t *set;
 } bitset;
 
-static void bitset_setat(bitset *bs, int index, int value);
+static void bitset_setat(bitset *bs, uint32_t index, uint32_t value);
 
-bitset *new_bitset(int limit);
-void bitset_insert(bitset *bs, int index);
-void bitset_remove(bitset *bs, int index);
-int bitset_contain(bitset *bs, int index);
+bitset *new_bitset(uint32_t limit);
+void bitset_insert(bitset *bs, uint32_t index);
+void bitset_remove(bitset *bs, uint32_t index);
+uint64_t bitset_contain(bitset *bs, uint32_t index);
 int bitset_eq(bitset* b1, bitset* b2);
 bitset *bitset_clone(bitset *b);
 bitset *bitset_intersection(bitset *b1, bitset *b2);
